requirements/python2.txt
Django==1.11.16
django-filter==1.1.0
django-leaflet==0.24.0
gunicorn==19.9.0
psycopg2==2.7.5
pytz==2018.5
